requirements.txt
tensorflow-gpu
opencv-python